Created MainSection component
This component is a layout for every screen in the app.
It accepts any body, the idea is to provide components as body according 
to the current screen.

One body component has already been created, but its not ready, it works 
as a component for tests. The way the buttons will appear on each 
sections is still not defined